Refactor and template "higher city"
