+++ question_generator/questionService.js
@@ -1,4 +1,3 @@
-// user-service.js
 const express = require('express');
 const bodyParser = require('body-parser');
